+++ backend/typst_writer/src/colors.rs
@@ -1,41 +1,29 @@
-static TEXT_HIGHLIGHT_COLORED: &str = "#8A6AA8";
-static TEXT_HIGHLIGHT_BW: &str = "#949494";
-static SOLUTION_FILL_COLORED: &str = "#F7F6E9";
-static SOLUTION_FILL_BW: &str = "#F9F9F9";
-static PRIMARY_COLOR_COLORED: &str = "#B79DCF";
-static PRIMARY_COLOR_BW: &str = "#222222";
-static SECONDARY_COLOR_COLORED: &str = "#B79DCF";
-static SECONDARY_COLOR_BW: &str = "#222222";
-static TERTIARY_COLOR_COLORED: &str = "#B79DCF";
-static TERTIARY_COLOR_BW: &str = "#222222";
+use types::{
+    colors::HexColor,
+    pdf::{DocumentOptions, SolutionDecoration},
+};
 
-pub fn get_color_preamble(has_color: bool) -> String {
-    let colored;
-    let solution_background;
-    let primary;
-    let secondary;
-    let tertiary;
-    if has_color {
-        colored = TEXT_HIGHLIGHT_COLORED;
-        solution_background = SOLUTION_FILL_COLORED;
-        primary = PRIMARY_COLOR_COLORED;
-        secondary = SECONDARY_COLOR_COLORED;
-        tertiary = TERTIARY_COLOR_COLORED;
-    } else {
-        colored = TEXT_HIGHLIGHT_BW;
-        solution_background = SOLUTION_FILL_BW;
-        primary = PRIMARY_COLOR_BW;
-        secondary = SECONDARY_COLOR_BW;
-        tertiary = TERTIARY_COLOR_BW;
+/// Reads the document options and writes the color macros in the Typst preamble
+pub fn get_color_preamble(document_options: &DocumentOptions) -> String {
+    let solution_text = &document_options.solution_text_color;
+    let solution_color = match document_options.solution_decoration {
+        SolutionDecoration::Fill => &document_options.solution_fill_color,
+        SolutionDecoration::Border => &document_options.solution_border_color,
+        SolutionDecoration::None => &HexColor::white(),
     };
+
+    let has_color = document_options.color;
+    let primary = HexColor::default_primary(has_color);
+    let secondary = HexColor::default_secondary(has_color);
+    let tertiary = HexColor::default_tertiary(has_color);
 
     format!(
         r#"
 #let primary_color = rgb("{primary}")
-#let solution_color = rgb("{solution_background}")
+#let solution_color = rgb("{solution_color}")
 #let secondary_color = rgb("{secondary}")
 #let tertiary_color = rgb("{tertiary}")
-#let colored(x) = text(fill: rgb("{colored}"), $#x$)
+#let colored(x) = text(fill: rgb("{solution_text}"), $#x$)
 #let primary(x) = text(fill: primary_color, $#x$)
 #let secondary(x) = text(fill: secondary_color, $#x$)
 #let tertiary(x) = text(fill: tertiary_color, $#x$)